Make sure ROOT is found with ROOTConfig.cmake first
Using $ROOTSYS/cmake/ROOTConfig.cmake is the recommended way and
$ROOTSYS/etc/cmake should only be added to CMake search paths in case
the former fails. Even though loading of the legacy modules would not
break the Unpacker build, when Unpacker2 is used as a subproject of
j-pet-framework, loading the legacy modules would globally shadow
functions provided by ROOTConfig, breaking the framework build.